+++ Tests/DependencyInjection/Fixtures/php/html_sanitizer.php
@@ -4,6 +4,7 @@ $container->loadFromExtension('framework', [
     'html_sanitizer' => [
         'sanitizers' => [
             'custom' => [
+                'default_action' => 'allow',
                 'allow_safe_elements' => true,
                 'allow_static_elements' => true,
                 'allow_elements' => [